backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
qdrant-client==1.9.1
openai==1.51.2
httpx==0.27.2
pydantic-settings
openpyxl
